requirements.txt
faiss-cpu
groq
langchain-groq
PyPDF2
langchain
langchain_community
langchain-huggingface
python-dotenv
pypdf
pyMuPDF
flask
flask_cors
pdf2image
gunicorn
sentence-transformers
torch